+++ rx_tif_lib/rx_tif.c
@@ -471,13 +471,13 @@ int tif_load(SPageId *id, const char *filedir, const char *filename, int printMo
 				int start;
 				start = -(INT32)wakeupLen;
 				memset(_ThreadPar[i].buffer+start*lineLen,           0x00, wakeupLen * lineLen);
-				memset(_ThreadPar[i].buffer+pinfo->lengthPx*lineLen, 0x00, wakeupLen * lineLen);
+				memset(_ThreadPar[i].buffer+(UINT64)pinfo->lengthPx*lineLen, 0x00, wakeupLen * lineLen);
 				
 				if (wakeupOn)
 				{
 					start=(-(INT32)wakeupLen+i*WAKEUP_BAR_LEN);
 					memset(_ThreadPar[i].buffer+start*lineLen, 0xff, WAKEUP_BAR_LEN * lineLen);
-					memset(_ThreadPar[i].buffer+(pinfo->lengthPx+wakeupLen-(i+1)*WAKEUP_BAR_LEN)*lineLen, 0xff, WAKEUP_BAR_LEN * lineLen);															
+					memset(_ThreadPar[i].buffer+((UINT64)pinfo->lengthPx+wakeupLen-(i+1)*WAKEUP_BAR_LEN)*lineLen, 0xff, WAKEUP_BAR_LEN * lineLen);															
 				}
 			};				
 		}
